fix(Pages.Main.NavBar): fix after code review
