+++ packages/client/src/components/LinkButton.tsx
@@ -1,0 +1,7 @@
+import { ButtonProps } from "@mui/material";
+import { Link, LinkProps } from "react-router-dom";
+import { Button } from "./Button";
+
+export function LinkButton(props: ButtonProps & LinkProps) {
+  return <Button component={Link} {...props} />
+}
